Move file, document and internal property types
