+++ src/backend/TrafficCourts/Citizen.Service/Services/ICitizenDocumentService.cs
@@ -1,5 +1,4 @@
-﻿using TrafficCourts.Common.Models;
-using TrafficCourts.Domain.Models;
+﻿using TrafficCourts.Domain.Models;
 using TrafficCourts.Coms.Client;
 
 namespace TrafficCourts.Citizen.Service.Services;
