+++ src/backend/TrafficCourts/Staff.Service/Services/DisputeService.cs
@@ -4,7 +4,6 @@ using System.Security.Claims;
 using System.Text.Json;
 using TrafficCourts.Common.Features.Lookups;
 using TrafficCourts.Common.Features.Mail.Templates;
-using TrafficCourts.Common.Models;
 using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using TrafficCourts.Coms.Client;
 using TrafficCourts.Messaging.MessageContracts;
@@ -74,7 +73,7 @@ public class DisputeService : IDisputeService
         if (dispute.NoticeOfDisputeGuid is not null && Guid.TryParse(dispute.NoticeOfDisputeGuid, out Guid noticeOfDisputeId))
         {
             // create new search properties
-            DocumentProperties properties = new DocumentProperties { NoticeOfDisputeId = noticeOfDisputeId };
+            Domain.Models.DocumentProperties properties = new Domain.Models.DocumentProperties { NoticeOfDisputeId = noticeOfDisputeId };
             disputeFiles = await _documentService.FindFilesAsync(properties, cancellationToken);
         }
 
@@ -85,7 +84,7 @@ public class DisputeService : IDisputeService
 
     private async Task<OcrViolationTicket?> GetOcrResultsAsync(Dispute dispute, CancellationToken cancellationToken)
     {
-        Coms.Client.File? file = await GetFileAsync(dispute, InternalFileProperties.DocumentTypes.OcrResult, cancellationToken);
+        Coms.Client.File? file = await GetFileAsync(dispute, Domain.Models.InternalFileProperties.DocumentTypes.OcrResult, cancellationToken);
         if (file is null)
         {
             return null;
@@ -104,7 +103,7 @@ public class DisputeService : IDisputeService
     /// <returns></returns>
     private async Task<Domain.Models.ViolationTicketImage?> GetViolationTicketImageAsync(Dispute dispute, CancellationToken cancellationToken)
     {
-        Coms.Client.File? file = await GetFileAsync(dispute, InternalFileProperties.DocumentTypes.TicketImage, cancellationToken);
+        Coms.Client.File? file = await GetFileAsync(dispute, Domain.Models.InternalFileProperties.DocumentTypes.TicketImage, cancellationToken);
         if (file is null)
         {
             return null;
@@ -131,7 +130,7 @@ public class DisputeService : IDisputeService
             return null;
         }
 
-        InternalFileProperties properties = new InternalFileProperties
+        Domain.Models.InternalFileProperties properties = new()
         {
             NoticeOfDisputeId = noticeOfDisputeId,
             DocumentType = documentType
