+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/FormRecognizerValidatorTest.cs
@@ -2,7 +2,6 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using TrafficCourts.Citizen.Service.Validators;
 using TrafficCourts.Common.Features.Lookups;
-using TrafficCourts.Common.Models;
 using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using Xunit;
