+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/Rules/CountSectionRuleTest.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using TrafficCourts.Citizen.Service.Validators;
 using TrafficCourts.Citizen.Service.Validators.Rules;
 using TrafficCourts.Common.Features.Lookups;
-using TrafficCourts.Common.Models;
 using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using Xunit;
 using static TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0.OcrViolationTicket;
